+++ android/src/main/java/com/brentvatne/common/api/Source.kt
@@ -52,16 +52,10 @@ class Source {
      */
     var drmProps: DRMProps? = null
 
-    /**
-     * Type of view to be used
-     */
-    @ViewType.ViewType
-    var viewType = ViewType.VIEW_TYPE_SURFACE
-
-    /** enable chunckless preparation for HLS
+    /** enable chunkless preparation for HLS
      * see:
      */
-    var textTracksAllowChuncklessPreparation: Boolean = false
+    var textTracksAllowChunklessPreparation: Boolean = false
 
     override fun hashCode(): Int = Objects.hash(uriString, uri, startPositionMs, cropStartMs, cropEndMs, extension, metadata, headers)
 
@@ -74,7 +68,6 @@ class Source {
                 cropEndMs == other.cropEndMs &&
                 startPositionMs == other.startPositionMs &&
                 extension == other.extension &&
-                viewType == other.viewType &&
                 drmProps == other.drmProps
             )
     }
@@ -137,9 +130,8 @@ class Source {
         private const val PROP_SRC_TYPE = "type"
         private const val PROP_SRC_METADATA = "metadata"
         private const val PROP_SRC_HEADERS = "requestHeaders"
-        private const val PROP_SRC_VIEW_TYPE = "viewType"
         private const val PROP_SRC_DRM = "drm"
-        private const val PROP_SRC_TEXT_TRACKS_ALLOW_CHUNCKLESS_PREPARATION = "textTracksAllowChunklessPreparation"
+        private const val PROP_SRC_TEXT_TRACKS_ALLOW_CHUNKLESS_PREPARATION = "textTracksAllowChunklessPreparation"
 
         @SuppressLint("DiscouragedApi")
         private fun getUriFromAssetId(context: Context, uriString: String): Uri? {
@@ -196,9 +188,8 @@ class Source {
                 source.cropStartMs = safeGetInt(src, PROP_SRC_CROP_START, -1)
                 source.cropEndMs = safeGetInt(src, PROP_SRC_CROP_END, -1)
                 source.extension = safeGetString(src, PROP_SRC_TYPE, null)
-                source.viewType = safeGetInt(src, PROP_SRC_VIEW_TYPE, ViewType.VIEW_TYPE_SURFACE)
                 source.drmProps = parse(safeGetMap(src, PROP_SRC_DRM))
-                source.textTracksAllowChuncklessPreparation = safeGetBool(src, PROP_SRC_TEXT_TRACKS_ALLOW_CHUNCKLESS_PREPARATION, true)
+                source.textTracksAllowChunklessPreparation = safeGetBool(src, PROP_SRC_TEXT_TRACKS_ALLOW_CHUNKLESS_PREPARATION, true)
 
                 val propSrcHeadersArray = safeGetArray(src, PROP_SRC_HEADERS)
                 if (propSrcHeadersArray != null) {