+++ android/src/main/java/com/brentvatne/exoplayer/ReactExoplayerViewManager.java
@@ -21,6 +21,7 @@ import com.brentvatne.common.api.ViewType;
 import com.brentvatne.common.react.VideoEventEmitter;
 import com.brentvatne.common.toolbox.DebugLog;
 import com.brentvatne.common.toolbox.ReactBridgeUtils;
+import com.brentvatne.react.ReactNativeVideoManager;
 import com.facebook.react.bridge.ReadableArray;
 import com.facebook.react.bridge.ReadableMap;
 import com.facebook.react.common.MapBuilder;
@@ -65,6 +66,7 @@ public class ReactExoplayerViewManager extends ViewGroupManager<ReactExoplayerVi
     private static final String PROP_DISABLE_DISCONNECT_ERROR = "disableDisconnectError";
     private static final String PROP_FOCUSABLE = "focusable";
     private static final String PROP_FULLSCREEN = "fullscreen";
+    private static final String PROP_VIEW_TYPE = "viewType";
     private static final String PROP_SELECTED_VIDEO_TRACK = "selectedVideoTrack";
     private static final String PROP_SELECTED_VIDEO_TRACK_TYPE = "type";
     private static final String PROP_SELECTED_VIDEO_TRACK_VALUE = "value";
@@ -91,12 +93,14 @@ public class ReactExoplayerViewManager extends ViewGroupManager<ReactExoplayerVi
     @NonNull
     @Override
     protected ReactExoplayerView createViewInstance(@NonNull ThemedReactContext themedReactContext) {
+        ReactNativeVideoManager.Companion.getInstance().registerView(this);
         return new ReactExoplayerView(themedReactContext, config);
     }
 
     @Override
     public void onDropViewInstance(ReactExoplayerView view) {
         view.cleanUpResources();
+        ReactNativeVideoManager.Companion.getInstance().unregisterView(this);
     }
 
     @Override
@@ -283,6 +287,11 @@ public class ReactExoplayerViewManager extends ViewGroupManager<ReactExoplayerVi
     @ReactProp(name = PROP_FULLSCREEN, defaultBoolean = false)
     public void setFullscreen(final ReactExoplayerView videoView, final boolean fullscreen) {
         videoView.setFullscreen(fullscreen);
+    }
+
+    @ReactProp(name = PROP_VIEW_TYPE, defaultInt = ViewType.VIEW_TYPE_SURFACE)
+    public void setViewType(final ReactExoplayerView videoView, final int viewType) {
+        videoView.setViewType(viewType);
     }
 
     @ReactProp(name = PROP_HIDE_SHUTTER_VIEW, defaultBoolean = false)